requirements.txt
tornado
numpy
pillow
requests
argparse
pyyaml
ruamel.yaml
pyserial
pyinstaller
